:recycle: api: move completion job to certif domain
