+++ api/tests/certification/evaluation/integration/infrastructure/repositories/jobs/certification-completed-job-repository_test.js
@@ -1,8 +1,8 @@
-import { certificationCompletedJobRepository } from '../../../../../lib/infrastructure/repositories/jobs/certification-completed-job-repository.js';
-import { CertificationCompletedJob } from '../../../../../src/certification/evaluation/domain/events/CertificationCompleted.js';
-import { LOCALE } from '../../../../../src/shared/domain/constants.js';
-import { JobPriority } from '../../../../../src/shared/infrastructure/repositories/jobs/job-repository.js';
-import { expect } from '../../../../test-helper.js';
+import { CertificationCompletedJob } from '../../../../../../../src/certification/evaluation/domain/events/CertificationCompleted.js';
+import { certificationCompletedJobRepository } from '../../../../../../../src/certification/evaluation/infrastructure/repositories/jobs/certification-completed-job-repository.js';
+import { LOCALE } from '../../../../../../../src/shared/domain/constants.js';
+import { JobPriority } from '../../../../../../../src/shared/infrastructure/repositories/jobs/job-repository.js';
+import { expect } from '../../../../../../test-helper.js';
 
 describe('Integration | Repository | Jobs | CertificationCompletedJobRepository', function () {
   describe('#performAsync', function () {
